Add image drawing to web back-end
This implementation uses a scratch HTML Canvas to hold the image. That's
most compatible when the source is an RGBA buffer and for general
transform/clip, but direct ImageData could be more efficient for
unity scale and clip to hard rectangle, and ImageBitmap could be useful
for letting the browser do the decoding (though ImageBitmap has poorer
compatibility and is async).

Also, alpha is interpreted as separate, not matching the doc comment or
the other back-ends.