Text trait refactoring
Make a new Factory type (with trait) so text layout doesn't require a
RenderContext. There's a `factory` method so if you have a RenderContext
you can do layouts from that, but it's anticipated there will be other
contexts (widget layout) where you want to do text layout but aren't
necessarily drawing to a surface.

WIP; as of this version, only the d2d back-end works.