Expose associated types
This patch exposes the backend-specific associated types for brushes,
text layout, and images.

It has similar goals as #37, but actually exposes all types as concrete
types, not just as type aliases for associated types. When these types
are hidden (as was the case for `Brush` in the d2d backend) then doing
an impl of a trait on the type causes coherence problems.

Also some updates to the documentation, making sure the new types (as
well as the `Piet` type alias) show up in documentation.