Remove `Result` when creating Font and TextLayout builders
Creating Font and TextLayout builders should not fail, so to avoid
unnecessary unwrapping for users, this commit removes the Result from
the `new_font_by_name` and `new_text_layout` methods in the `Text`
trait.

This is a user API breaking change.

Also updated implementers of `Text` (all existing backends):
- null_renderer
- direct2d
- cairo
- web

Also updated user of `Text`, `piet-crate`.

As an ergonomic change for devs, this commit also:
- adds `#![cfg(windows)]` to the `direct-2d` backend so that it does
not attempt to compile on linux.
- adds `.cargo` to .gitignore; it's used for cross compilation config