[cairo] first pass hit_test_point, w/ binary search
- helper fn for getting the boundaries of the codepoint for any selected
byte index.

General algorithm for checking if point_x is within a codepoint's
boundaries:
- check if outside line. If so, return first or last text position, but
saying it's an outside hit (need to check windows api for compatibility)
- check if it's start or end char (probably most common clicks)
- binary search. It's a loop, and currently assumes that a hit must be
reached (since the above check should guaranteed that it's in the text).

Need to ask whether a text position needs to be char index or can be
byte index.

Does not yet handle lines