Fix closed paths on stroke (d2d)
Do a slightly more inefficient conversion to bezier paths so we can
tell in advance whether a subpath needs to be closed. This is a bit of
a workaround, as there are plans to improve the direct2d bindings so
we can set the is_closed at the end of the subpath, enabling a single
pass of iteration (tracked in #65).