Remove refs from WebRenderContext constructor API
Specifically, the lifetime of WebRenderContext is changed to not depend on lifetimes of CanvasRenderingContext2d and Window objects, since these lifetimes don't actually enforce the existence of web resources anyways.

The mut is also removed from CanvasRenderingContext2d since its API has
no mut calls and this generally makes the WebRenderContext difficult to use
in libraries like druid.

For compatibility the WebRenderContext maintains a phantom lifetime
placeholder for compatibility as well as for anticipating a correct
lifetime there.