Fix piet-web tests & add console hook for panics
This commit tweaks the tests to pass, and adds a useful console log hook to capture panics in the browser console.

Admittedly, the tweak in tests is perhaps dependent on my system configuration, so it will be worth for reviewers
to verify whether the tests pass. To do so please check the console log in the browser.