Final tweaks to integrate piet-coregraphics
This involves explicitly handling the different coordinate spaces
between piet and coregraphics. I've chosen to be explicit; when
creating a CoreGraphicsContext, you must either ask for y-up and pass
in a height (so that we can apply a transform under the hood) or you
can ask for y-down and handle conversion yourself (which plays nicely
with Cocoa, and NSView.isFlipped.