[cg] Fix image drawing
Images were not drawing correctly if the rect they were drawn into
was not at the origin.

This is related to coregraphics coordinate space wonkiness; the
easiest solution I see is to transform the context and then draw
the image at the origin, instead of bothering with correctly
transforming the rect.