Support default attributes on TextLayout types
This adds the ability to set the default attributes of various kinds when
constructing a text layout. These attributes will be used when styling
text that does not have explicit attributes applied.

With this change, font size is now ignored; updating the font types
to clarify this will be an upcoming patch.

This also renames `add_attribute` to `range_attribute`, which
better matches the new `default_attribute` method name.