Compute default line height for empty layout
This attempts to address the issue of how to determine the correct
cursor size in an empty layout.

Essentially: when constructing the layout, we compute the line height
and baseline values that would be used for the layout's default
font and font size, and we stash these; if the layout is empty, we
use these values to synthesize the result of calls to hit_test_position
and line_metric.

- closes #268