Add TextStorage trait
This is a further refinement of how we want to store the
actual text backing a layout object. The motivation here
is to allow the client maximum flexibility in terms of
how they store their text.

This moves from using `Arc<str>` everywhere to using
`Rc<dyn TextStorage>`; this introduces another layer
of pointer indirection in various cases, but saves us
needing to add a type parameter.