Add conformance tests to piet-common
This is an experiment? but it would be nice if we could
smoke test the behaviour of the main backends, to check
big-picture behaviour.

This *might* not be the best place for this; in particular
it means that backends that aren't in piet-common won't
have these tests, but I'm not *practically* concerned
about that.

Back in the mists of time we had a piet-test crate that
was intended to do what this is doing, but it was folded
into the main piet crate. I think another viable approach
would be to define individual testing functions in piet,
and then have the various backends call into piet to run
them, similar to how the sample pictures work, but for
the time being this feels like a simple way to do what
we would like.