Fix panic in line_range with empty string
This was a mess in a number of ways, and would crash if called
when the layout had no lines.

The expected behaviour here is debatable; we could return the
empty string or we could return None. I chose to return the
empty string to match the behaviour of Layout::line_metric,
which still returns a metric when there are no lines (in that
case, so that we can draw the cursor in an empty layout).