[cg] Compute LineMetrics greedily
Unlike the other backends, with coretext we were not storing
a Vec<LineMetric> in the layout object, but instead generating
them on demand.

As I work on standardizing the behaviour around newlines at EOF
and treating an empty layout as having a single line, it is easier
to reason about a single source of 'line truth'; the logic started
to get hard to follow in the previous case.

This also cleans up the API in ct_helpers somewhat; much of it
was written before some improvements were upstreamed to the
core-text crate, and we can be less fancy there.

This includes one small change to behaviour: we were previously
storing the floored baseline in the line_y_positions vector,
which I _think_ was hiding a small bug. This may cause subtle
changes to rendering behaviour, which I will correct in snapshots
if required.