[d2d] Use default line height as height when text is empty
This is a less principled approach to fixing #339 than is
outlined in that issue, but is less invasive and works just as well.

This also fixes a test in piet-common that was intended to catch
this, but that was using a layout with a default font size, which
doesn't manifest this problem.

- fixes #339