Add Color::from_hex_str for parsing CSS-style colors
This code is a bit idiosyncratic because I wanted it to
work in const contexts.

Question: should we have a `FromStr` impl? This might make sense,
but it's not 100% clear since there are any number of ways to
parse a string into a color. Would this cause problems in practice?

- closes #348