Update cairo/web to xi-unicode 0.3
We had alread bumped web and not noticed the breakage because
we don't actually run any tests; we'd noticed the breakage in
cairo and not bumped.

This standardizes on 0.3 and fixes the breakage by manually
addung the last line if there's no explicit linebreak.

This should address https://github.com/linebender/druid/issues/1488