[cg] Fix incorrect line ranges with trailing tabs
The CTLineGetStringRange function I was using for string
ranges apparently doesn't include the range of all trailing
whitespace. This seems to maybe only be a problem for tabs?
In any case, it means we were getting subtly wrong reported
line ranges, which I bumped into with IME.

This fixes that, but just using the start offsets of lines;
we assume a line is the range between two line starts.