[cg] Fix incorrect layout of text containing spaces
tl;dr: Don't use CTFramesetterSuggestFrameSizeWithConstraints
to determime the size of your frame.

This was suggested to us in https://github.com/linebender/piet/issues/413

A consequence of this change is that we need to calculate our
actual line positions ourselves, but this ends up not being too
tricky.

This breaks a few tests that were counting on the CoreText's
line-spacing secret sauce that shows up in fonts like Helvetica;
I am making the decision to not try and replicate that behaviour,
and instead defer to the font metrics.

Fairly miraculously, with this patch we are managing to very closely
match the drawing behaviour of CTFrame in most cases.

There are a few broken snapshots; this is the result of us now
doing floor(ascent + 0.5) there; our previous behaviour was
incorrect.