Rework the StrokeStyle type
This started out as my just wanting to fix #416 while it was
on my mind, and then, well, things got a little out of hand.

I realized I wasn't familiar with some of the terminology used,
so I patched up the docs based on the postscript manual. Then
I noticed there were a few areas where the API seemed confusing,
and where the backends had to duplicate a bunch of code in order
to set defaults, so I went through and implemented `Default`,
using the defaults mentioned in the spec.

Highlights:

- 'dash' is replaced with separate 'dash_pattern' and 'dash_offset'
  methods
- StrokeStyle can now be constructed in a const context
- dash pattern can now be represented with a static &[f64] slice
- miter limit is now a member of the LineJoin::Miter variant.