Clarify and test behaviour of hit_test_position
This was motivated by the pango work; I realized that the
behaviour in the case of interior bytes was not well defined.

This ensures checks to ensure backends panic when the provided
index is not a codepoint boundary, and notes that behaviour
may vary if the index is not a grapheme boundary, because
this is not a case I feel compelled to support.