Fix `capture_image_area` not respecting CG transformation matrix
The `capture_image_area` method of the CoreGraphicsContext was not
taking into account that a transformation matrix is applied to the
context. As a result, it only worked correctly if one applied this
scaling manually to the `src_rect` prior to passing it to the method.

With this addition, we now respect the CG transformation matrix and
apply it to the `src_rect` prior to cropping the captured `CGImage`.