Initial import of cpu-sparse prototype
This copies the implementation from the cpu-sparse branch of compute-shader-101, which is currently the most up-to-date version (though there are still rendering flaws).

We remove the bytemuck dependency and change colors the color crate (following tip of tree peniko), but otherwise unchanged. Lots of warnings as well, as basically everything is unused.

There will be changes. In particular, path_id will be removed, as there will be a separate sort per path.